refactor: 로직을 모듈로 분리
