+++ scripts/utils/constant.js
@@ -1,0 +1,4 @@
+const EXIST_EMAIL = "test@codeit.com";
+const EXIST_PASSWORD = "codeit101";
+
+export { EXIST_EMAIL, EXIST_PASSWORD };
